Fix mrr type lie in listCompanies (Task 12 reviewer finding)
COALESCE(s.mrr_amount, 0) AS mrr returned a stringified NUMERIC over
JSON despite CompanySummary.mrr being typed number. metricsOverview's
current_mrr already cast ::float; apply the same cast here so the
type is honest for any future arithmetic/sorting on mrr.

+++ backend/platform/api.ts
@@ -85,7 +85,7 @@ export const listCompanies = api(
 
     const listSql = `
       SELECT c.id, c.name, c.industry, c.plan_tier, c.customer_stage,
-        COALESCE(s.mrr_amount, 0) AS mrr
+        COALESCE(s.mrr_amount, 0)::float AS mrr
       FROM companies c
       LEFT JOIN subscriptions s ON s.company_id = c.id
       ${whereClause}
